requirements.txt
fastapi
uvicorn
groq
google-generativeai
supabase
python-dotenv
resend
pydantic
pydantic-settings
python-multipart
postgrest
